[netboot] Consistently namespace under "netboot"
This change is a pure rename: apart for some minor name tweaks for
clarity, this change ensures that all netboot definitions are prefixed
with either "netboot_" or "NETBOOT_". This is an established convention
for C libraries - and will be what we will want out of C bindings once
we FIDL-ify the protocol.

Bug: 91102
Change-Id: I9433bb4fef50b727d59d8e344a9679ce99032990
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821136
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>