[zither][rust] Annotate bindings with derived traits
These basic traits enable conveniences taken for granted, and thus are
prerequisites for generating useful bindings. The default traits derived
for everything are Copy, Clone, Debug, Eq, and PartialEq. Additionally,
zerocopy::{AsBytes, FromBytes} are derived where supported, which
streamline conversion to and from byte slices.

AsBytes and FromBytes require knowledge about the whole layout in order
to determine support. So that (a) this does not introduce logic to
reprocess each declaration's 'dependencies' in the generation of its
binding, and that (b) we do not complicate the main zither summarization
logic with backend-specific bookkeeping, we introduce a scheme so that
backends can provide a callback to be passed to zither.Summarize for
additional bookkeeping.

Bug: 91102
Change-Id: I118fb7556729b8d0bacbfeab0c12f2f68c341f27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822048
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>