[audio] virtual_audio_util: add usage and README
This CL adds to the virtual_audio CLI utility a --help flag that
details its various cmdline options, and a README that answers
common questions.

The utility now explicitly checks whether a specified initial
clock-rate is out-of-range, and fails in that case. (Previously
it did this only for rate-adjustments made once the device was
active.)

Small adjustments are also made to the available format and
gain options, and the default '--write-rb' value.

The CL also includes a number of 'lint'-type fixes from clang.

Test: diagnostic CLI utility for test fixture, no product change

Change-Id: Ied10b056e6ec402424451f80eeb1796ec4b79b80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827496
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>