[ums] Simplify and re-eanble ums-integration-test.
This CL removes the Device protocol in
fuchsia.hardware.usb.peripheral.block. The protocol isn't actually used
to enable/disable the writeback cache in the ums-function driver (there
is no writeback cache). More importantly, the ums-function driver is
used for integration testing of the usb-mass-storage driver, and the
protocol is superfluous in that regard.

The test case UmsTest.CachedWriteWithNoFlushShouldBeDiscarded is flawed
in that it actually verifies that the supposedly cached write is not
discarded. This CL removes the test case.

Bug: 121731
Test: fx test ums-integration-test
Change-Id: Icd8c06bda26404017f0c33a05d2b4d1d6d3be2be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/825071
API-Review: Christopher Anderson <cja@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>