[intel-i915] Descriptions for relevant Memory Controller registers.
The on-processor Memory Controller, also known as the MCH (Memory
Controller Hub), has its PCI BAR aperture mirrored as read-only in the
graphics BAR space. This is intended to make it easy for the display
driver to learn low-level DRAM configuration details needed for
memory bandwidth computations, and for configuring the arbiter.

This CL adds definitions for the MCH registers that are relevant to
display engine configuration.

Bug: 111612
Change-Id: I33c8fd6a669a5a9527646c0d9451717b2ff19828
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792122
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>