[ffx] Make RCS state less dependent on host pipe
In preparation for USB-connected targets, which don't have a host pipe
associated with them, we need to teach FFX to not equate presence of a
host pipe with connectivity.

This code changes the host pipe to check for an existing RCS connection,
and if that connection is working, to end immediately. The connectivity
check has a 1 second timeout but will usually fail faster than that.
Possibly we could see some odd behavior on devices which are running
very slowly if RCS takes more than a second to respond to a knock
request.

Most code that checks for the presence of a host pipe explicitly only
checks if there is a handle to a host pipe task, not what the state of
that task is, so this will make USB-connected targets appear to have a
host pipe as far as the rest of the target collection code is concerned.

Bug: 124481
Change-Id: Ib44c79a3b7827484a6621991ab49080ff790e6e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827083
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>