[gigaboot] Add an indication that boot mode skips normal boot
When trying to boot Fuchsia on a NUC I've ran into an occasion several
times where the NUC refuses to boot and goes straight into fastboot. I
was only able to recover this by entering the UEFI shell and running
`dmpstore -d bootbyte`, to delete the boot mode variable.

I'm not familiar with the mechanics of this variable. This CL adds a
note during the boot process when that is the case.

Change-Id: I18ce45a2e96c2a6498bf8b696cd18281bdad9a6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829642
Reviewed-by: David Pursell <dpursell@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>