[reachability] Improve clippy linting for async
This is a minor bug, and according to
rust-lang.github.io/rust-clippy/master/index.html#unused_async it is
beneficial to check for this and remove it to reduce mental and
computational overhead with this declaration.

This CL adds a warn clippy tag for functions that hae an async tag
without having any awaits within the function body.

With this lint rule, two cases were caught and fixed:
- `resume` in NetworkChecker
- `handle_ten_secondly_telemetry` in Telemetry

Fix: 124592
Change-Id: Idf6f975493ca3377b47b9472b3852d7937a75bc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/826659
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>