[fxfs] Simplify checking of directory children
Prior to this change, fsck followed different paths depending on the
order in which objects were encountered.  This change makes it so that
directory children are processed in a second pass (the same pass that we
currently scan extents in), and allows for a few simplifications.

This CL also includes some other changes:

 * Directory::update_attributes now preserves the operation which
   is required to ensure object counts are correctly updated.  It has
   also been simplified.
 * Directory::insert_child will now update the sub_dirs field when
   inserting a directory.
 * The algorithm for checking link cycles has been changed so that it
   now catches the case where two directories refer to each other as
   children.  This change in algorithm means we no longer need to record
   all the children of directories (which should reduce memory usage).

Fixed: 124705
Change-Id: I3b5dc94ddccf72370ee070c97dae0205627728d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/829136
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>