[netstack3] Implement timeout for defunct FIN_WAIT_2 connections
FIN_WAIT_2 is a potential state that can stuck forever in the system:
if the peer end just dies, and the connection becomes defunct/orphaned,
there is no way that we will ever look into the connection; in other
'closing' states, we either are waiting for a `close` from the user or
we are waiting for an ACK, whose absence will timeout the connection.
To solve this problem of lingering FIN_WAIT_2 states, we need a timer
for the state that is not described in the RFC, but is de facto present
in most implementations.

Bug: 117956
Change-Id: I92cfcd4db124473dd44b7fa78806f20547342241
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827456
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>