[starnix] Implement the remote binder driver device.
The device expects a userspace process to connect to it and use an ioctl
to transmit the name of the service to use to start a remote fuchsia
component implementing using binder.

The driver will connect to the remote process using the RemoteController
protocol, and will drive the local process so that local task are
associated with remote threads when doing binder ioctls.

Change-Id: Ic5a9b22f94ec9b65563f8a6631f1f9ffe2b3b505
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/823787
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>