[mdns] answer instance-related questions using correct sections
When answering questions, InstanceResponder was always putting the
PTR resource in the answer section and everything else in the
additional section. This is correct for PTR queries, but not
for other queries.

This CL updates the InstanceResponder so that it puts the resource
matching the question in the answer section and other resources
in the additional section. When responding to ANY questions, the
PTR resource is the answer if the service type is specified, and
the SRV and TXT resources are the answer if the service instance
is specified.

Fixed: 104937

Test: new unit tests

Change-Id: I0c14eabd4c61ace9da19610e0501c461409665ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813967
Commit-Queue: Dale Sather <dalesat@google.com>
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Robert Quattlebaum <rquattle@google.com>