[codecs][sbc] Send queued audio when space is low
When space in the output buffer is low (less than enough to store
another packet output), send the queued output, instead of relying on
the output matching exactly the correct size.

Not doing this means the SBC decoder sends an error status because there
is not enough space in the output buffer (and drops the input frame).

Bug: 122271
Test: updated tests to add some buffer constraints
Test: fx test audio_decoder_test

Change-Id: I8e14d08880866a95fc0c47548ceeb1f02aeb0577
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816930
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>