[testrunner] Reduce SSH reconnect timeout.
Before, we automatically retried to connect if we lost SSH connection
during a test with a timeout of 2 minutes. This would allow us to
reconnect even if the target underwent a reboot. We should reduce the
timeout so that we can report failures due to the target rebooting so
that they can be investigated in case the reboot is caused by a certain
test.

Change-Id: I9a99e17bb6643c5e56cafa3c1069df50b03f4c1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830664
Reviewed-by: Andrés Villegas <andresvi@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>