[diagnostics][triage] Delete `Multiple` error
Triage-lib has a kind of error called "Multiple" that's a collection of
other errors, so every error can be seen and acted on.

In practice, this causes code clutter and also clutter in output. We
should treat an error more like a thrown exception (short-circuiting
evaluation), not assume that every error needs to be user-facing.

This CL deletes the Multiple error type entirely, dropping any errors
beyond the first.

It does not delete the "Ignore" error list - that may be a good idea,
but it's a separate issue.

Fixed: 106637
Bug: 58922
Change-Id: I46ebdbd3bb024d5e2acf44873f7435a6526f66fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830778
Commit-Queue: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>