[a11y] Wires a11y manager with new gd protocol.
This change is part of the accessibility migration to use the new gesture
disambiguation protocol.

this change instantiates a touch source to be used by the a11y manager, and
passes it to the gesture manager.

the Screen Reader and magnifier then register their gestures to process the
arriving pointer events.

Test: fx test a11y-lib-tests
Bug: 100030
Change-Id: I54cbc0b1291b9acf73d35b6449b1620018f353d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801599
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>