[netstack] Deflake Fast UDP unit test
TestDatagramSocketWithBlockingEndpoint/closeAfterUnblocked verifies that
when a socket is closed, it will finish writing any available packets
into the gVisor endpoint (as long as it is not blocked) before cleaning
up. The test, as written, is racy because it does the following:

1. Write packets into the zircon socket
2. Unblock the endpoint
3. Close the socket
4. Expect all packets to be enqueued

And it is possible for the write loop to observe (2) and (3)
simultaneously, in which case there is no guarantee it will finish
enqueueing packets. To deflake this, wait for the write loop to resume
enqueueing packets before initiating socket closure.

Fixed: 123927
Multiply: netstack-gotests
Change-Id: I4865b4947e3d34fb1b0abd89c361c0095e8fd01f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831917
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>