[ffx][config] Make Cache not depend on Environment
This makes the Cache type take the value it stores as a generic so that
the code in that file doesn't need to depend on environmentcontext or
environment at all. This removes a circular dependency within the config
crate that makes it hard to break up into smaller pieces.

Multiply: ffx_config_lib
Change-Id: I02ab3c0edfb2b8fbb622fb8ead012bd5e3db4984
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/828878
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>