[netstack3] Allow device routes with a gateway
Allow clients to add routes that specify both the device and a gateway.

Similar to adding a pure gateway route, first check that the gateway "Is
a neighbor" by verifying that there is an on-link route to the gateway.
However, unlike a pure gateway route, only consider routes over the
given device.

While I was here, I decided to consolidate all of our "add route"
functions into a single function, and make better use of the
`AddableEntry` type.

Fixed: 96680

Change-Id: I3ee79fcbbf4befbf247ddf8f8213d24dfbdf933c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/831516
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>