[driver_host] Release api lock when making sync calls
Holding the API lock when making blocking calls can result in a deadlock
as the driver manager may end up blocked on call back into the driver
host. Releasing the for the duration of the sync call should be safe as
long as the device isn't destroyed, which it shouldn't be as we are
hanging onto a reference.

Bug: b/276771439
Change-Id: I5080ec0621f2afe57572985f1265299418c57a4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832248
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>