[inspect] Fix bug in the `Heap/extend` benchmark
We wereusing `iter_with_large_setup` which internally calls
`iter_batched`. This function moves the input to the function
being measured which leads to measuring the Drop impl of the input.

The correct thing to do here is to use `iter_batched_ref` which
passes a reference to the function executing the code that will be
measured. That way the input is dropped after the function has been
measured and not during.

Bug: 123517

Change-Id: I911ca6fed95fac69a87c300dcca1c7cf37318663
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832276
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>