[fx] Add get-flash-source command.
This command is used to retrieve the path to either
a flash manifest, or to a bazel-generated product bundle
when `use_bazel_images_only` is set in the current build
configuration (though this is an implementation detail
that may change in the future).

The output string uses a prefix of either
`flash-manifest:` or `product-bundle:` to distinguish
between these two cases.

This will later be used by `fx flash` and `fx remote-flash`

Bug: 124241

Change-Id: I00747d0858e5d42caada878fb3e5ca52c8b647fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832500
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>