[chromium][ui][test] `use` ContextProvider only when necessary
Previously, `web_engine_base.shard.cml` included a `use` of
"fuchsia.web.ContextProvider", which doesn't necessarily always make
sense and is different from the other `use`s in that that file.
Remove this line from that file and add it to the tests that need it.
Those tests launch a chromium/web/client Component and offer a hermetic
ContextProvider.

Bug: 123550
Change-Id: I383077b592a2934a01d858683c648f8467e18b26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832697
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>