[fuchsia.io] Remove RightsRequest
This was an over-engineered solution to needing POSIX semantics on Open
calls that are not known a priori to be reaching a directory.

The same information is now in the directory variant of NodeProtocols,
which is now a table.

Bug: 77623
Change-Id: Ib8c5a5b2736377a42a5080468c54599bf0bcfcf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/832777
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Adam Barth <abarth@google.com>